Tools

Packing

Tools which ease the the weights (de)serialization.

Fed.Serde.pack — Function

pack(data)

Utility function to serialize data into a byte array.

Fed.Serde.unpack — Function

unpack(bytes)

Utility function to deserialize bytes into its original representation.

Quantizer

Tools to (de)quantize a vector of values.

Fed.Serde.Quantizer — Type

Quantizer{T <: Unsigned}

Apply value quantization by scaling it down from one type to another, which requires less bits than the original type.

Fed.Serde.Quantizer — Method

Quantizer{T}(minval, maxval)

Create a new Quantizer which can quantize values from the range [minval, maxval].

Fed.Serde.Quantizer — Method

Quantizer{T}(data::Vector{Float32})

Create a new Quantizer for the given data.

Fed.Serde.quantize — Function

quantize(q::Quantizer{T}, x) where T <: Unsigned

Quantize x with the Quantizer q.

Fed.Serde.dequantize — Function

dequantize(q::Quantizer, x)

Dequantize x with the Quantizer q.

Delta compressor

Expose a generic diff and patch mechanism to remove redundant elements from a Vector.

Fed.Serde.Patch — Type

Patch{T <: Any}

Data structure which contains the diff between two Vector. The second Vector can then be rebuilt by patching the first one with the Patch.

Fed.Serde.diff — Function

diff(old::Vector{T}, new::Vector{T}, sizeof_T::Int)

Generate a Patch by delta compressing the difference between old and new.

Fed.Serde.patch — Function

patch(old::Vector{T}, patch::Patch{T})

Apply patch to old.